[FIX] accuracy

+++ src/dataloader/sentence_classification_dataloader.py
@@ -28,6 +28,6 @@ class SentenceDataset(Dataset):
         return {
             'input_ids': encoding['input_ids'].flatten(),
             'attention_mask': encoding['attention_mask'].flatten(),
-            'label': torch.tensor(label, dtype=torch.long)
+            'label': torch.tensor(label, dtype=torch.float)
         }
